+++ EurofurenceTests/Application/Components/Dealers/Presenter Tests/Categories/WhenDealersSceneRevealsCategoryFilters_DealersPresenterShould.swift
@@ -1,56 +1,6 @@
 import Eurofurence
 import XCTest
 
-class FakeDealerCategoriesViewModel: DealerCategoriesViewModel {
-    
-    private let categories: [DealerCategoryViewModel]
-    
-    var numberOfCategories: Int {
-        return categories.count
-    }
-    
-    func categoryViewModel(at index: Int) -> DealerCategoryViewModel {
-        return categories[index]
-    }
-    
-    init(categories: [DealerCategoryViewModel] = []) {
-        self.categories = categories
-    }
-    
-}
-
-class FakeDealerCategoryViewModel: DealerCategoryViewModel {
-    
-    let title: String
-    private var isActive: Bool = false
-    
-    init(title: String) {
-        self.title = title
-    }
-    
-    func add(_ observer: DealerCategoryViewModelObserver) {
-        if isActive {
-            observer.categoryDidEnterActiveState()
-        } else {
-            observer.categoryDidEnterInactiveState()
-        }
-    }
-    
-    private(set) var toldToToggleActiveState = false
-    func toggleCategoryActiveState() {
-        toldToToggleActiveState = true
-    }
-    
-    func enterActiveState() {
-        isActive = true
-    }
-    
-    func enterInactiveState() {
-        isActive = false
-    }
-    
-}
-
 class WhenDealersSceneRevealsCategoryFilters_DealersPresenterShould: XCTestCase {
 
     func testBindTheAvailableCategoriesToTheScene() {
